GenProjects can now be run from anywhere + buffer element offset warning fix

+++ Hazel/src/Hazel/Renderer/Buffer.cpp
@@ -61,7 +61,7 @@ BufferLayout::BufferLayout(const std::initializer_list<BufferElement>& elements)
 
 void BufferLayout::CalculateOffsetsAndStride() 
 {
-	uint32_t offset = 0;
+	size_t offset = 0;
 	m_Stride = 0;
 
 	for (auto& element : m_Elements)
